refactor: streamline agent and model runner integration, removing unused code and enhancing clarity
