+++ src/bub/builtin/agent.py
@@ -13,27 +13,21 @@ from dataclasses import dataclass, replace
 from datetime import UTC, datetime
 from functools import cached_property
 from pathlib import Path
-from typing import Any, Literal, cast
+from typing import Any
 
-from any_llm import AnyLLM
-from any_llm.types.completion import (
-    ChatCompletion,
-    ChatCompletionChunk,
-    ChatCompletionMessage,
-    ChatCompletionMessageFunctionToolCall,
-    ChatCompletionMessageToolCall,
-    ChoiceDeltaToolCall,
-    Function,
-    ParsedChatCompletion,
-)
 from loguru import logger
-from pydantic import TypeAdapter, ValidationError
 
-from bub.builtin.settings import ModelCandidate, load_settings
+from bub.builtin.model_runner import (
+    ModelOutputAccumulator,
+    ModelRunner,
+    is_context_length_error,
+    tool_invocation_from_native,
+)
+from bub.builtin.settings import load_settings
 from bub.builtin.store import ForkTapeStore
 from bub.builtin.tape import TapeService
 from bub.framework import BubFramework
-from bub.runtime import AsyncStreamEvents, BubError, ErrorKind, StreamEvent, StreamState
+from bub.runtime import AsyncStreamEvents, BubError, StreamEvent, StreamState
 from bub.skills import discover_skills, render_skills_prompt
 from bub.tape import InMemoryTapeStore, Tape
 from bub.tools import (
@@ -47,12 +41,7 @@ from bub.utils import workspace_from_state
 
 CONTINUE_PROMPT = "Continue the task until all targets are completed."
 HINT_RE = re.compile(r"\$([A-Za-z0-9_.-]+)")
-_CONTEXT_LENGTH_PATTERNS = re.compile(
-    r"context.{0,20}(?:length|window)|maximum.{0,20}context|token.{0,10}limit|prompt.{0,10}too long|tokens? > \d+ maximum",
-    re.IGNORECASE,
-)
 MAX_AUTO_HANDOFF_RETRIES = 1
-TOOL_ARGUMENTS_ADAPTER = TypeAdapter(dict[str, Any])
 
 
 class Agent:
@@ -71,6 +60,10 @@ class Agent:
             tape_store = InMemoryTapeStore()
         tape_store = ForkTapeStore(tape_store)
         return TapeService(bub.home / "tapes", tape_store, self.framework.build_tape_context())
+
+    @cached_property
+    def model_runner(self) -> ModelRunner:
+        return ModelRunner(self.settings)
 
     @staticmethod
     def _events_from_iterable(iterable: Iterable) -> AsyncStreamEvents:
@@ -249,7 +242,7 @@ class Agent:
             except Exception as exc:
                 error_message = f"{exc!s}"
                 elapsed_ms = int((time.monotonic() - start) * 1000)
-                if auto_handoff_remaining > 0 and _is_context_length_error(error_message):
+                if auto_handoff_remaining > 0 and is_context_length_error(error_message):
                     auto_handoff_remaining -= 1
                     logger.warning(
                         "auto_handoff: context length exceeded, performing automatic handoff. tape={} step={}",
@@ -396,33 +389,23 @@ class Agent:
             from bub.builtin.tools import model_tools
 
             model_tools_for_call = model_tools(tools)
-            text_parts: list[str] = []
-            tool_calls = _ToolCallAccumulator()
-            response: ChatCompletion | ParsedChatCompletion[Any] | None = None
+            output = ModelOutputAccumulator()
             async with asyncio.timeout(self.settings.model_timeout_seconds):
-                completion = await self._completion_response(
+                async for event in self.model_runner.run(
                     model=model or self.settings.model,
                     messages=messages,
                     tools=model_tools_for_call,
-                )
-                if isinstance(completion, ChatCompletion):
-                    response = completion
-                async for event in _completion_events(completion, state, text_parts, tool_calls):
+                    state=state,
+                    output=output,
+                ):
                     yield event
 
-            assistant_message = response.choices[0].message if response is not None else None
-            text = (
-                assistant_message.content
-                if assistant_message and assistant_message.content is not None
-                else "".join(text_parts)
-            )
-            native_tool_calls = tool_calls.as_native()
-            if native_tool_calls:
+            text = output.text
+            tool_calls = output.tool_calls
+            if tool_calls:
                 tool_map = {tool_item.name: tool_item for tool_item in model_tools_for_call}
-                serialized_tool_calls = [tool_call.model_dump(exclude_none=True) for tool_call in native_tool_calls]
-                tool_invocations = [
-                    _tool_invocation_from_native(tool_call, tool_map) for tool_call in native_tool_calls
-                ]
+                serialized_tool_calls = [tool_call.model_dump(exclude_none=True) for tool_call in tool_calls]
+                tool_invocations = [tool_invocation_from_native(tool_call, tool_map) for tool_call in tool_calls]
                 yield StreamEvent("tool_call", {"tool_calls": serialized_tool_calls})
                 context = ToolContext(tape=tape.name, run_id=run_id, state=tape.context.state)
                 execution = await ToolExecutor().execute_async(
@@ -437,7 +420,7 @@ class Agent:
                     response_text=None,
                     tool_calls=serialized_tool_calls,
                     tool_results=execution.tool_results,
-                    response=response,
+                    response=output.response,
                     model=model or self.settings.model,
                     usage=state.usage,
                 )
@@ -453,44 +436,13 @@ class Agent:
                 system_prompt=system_prompt,
                 new_messages=[prompt_message],
                 response_text=text,
-                response=response,
+                response=output.response,
                 model=model or self.settings.model,
                 usage=state.usage,
             )
             yield StreamEvent("final", {"ok": True, "text": text})
 
         return AsyncStreamEvents(iterator(), state=state)
-
-    def _build_llm(self, candidate: ModelCandidate) -> AnyLLM:
-        return AnyLLM.create(
-            candidate.provider,
-            **self.settings.model_client_kwargs(candidate.provider),
-        )
-
-    async def _completion_response(
-        self, *, model: str, messages: list[dict[str, Any]], tools: list[Tool]
-    ) -> ChatCompletion | ParsedChatCompletion[Any] | AsyncIterator[ChatCompletionChunk]:
-        from bub.builtin.tools import completion_tools
-
-        tool_payloads = completion_tools(tools) or None
-        completion_messages: list[dict[str, Any] | ChatCompletionMessage] = list(messages)
-        candidates = self.settings.model_candidates(model)
-        for index, candidate in enumerate(candidates):
-            try:
-                llm = self._build_llm(candidate)
-                return await llm.acompletion(
-                    model=candidate.model_id,
-                    messages=completion_messages,
-                    tools=tool_payloads,
-                    max_tokens=self.settings.max_tokens,
-                    stream=llm.SUPPORTS_COMPLETION_STREAMING,
-                )
-            except Exception as exc:
-                if index == len(candidates) - 1:
-                    raise
-                logger.warning("model candidate failed; trying fallback model={} error={}", candidate.name, exc)
-
-        raise RuntimeError("no model candidates available")
 
     def _system_prompt(
         self, prompt: str, state: State, allowed_skills: set[str] | None = None, tools: Iterable[Tool] | None = None
@@ -512,131 +464,6 @@ class Agent:
         if "context" in tape.context.state:
             return f"{CONTINUE_PROMPT} [context: {tape.context.state['context']}]"
         return CONTINUE_PROMPT
-
-
-@dataclass
-class _StreamToolCall:
-    id: str | None = None
-    type: Literal["function"] | None = None
-    name: str | None = None
-    arguments: str = ""
-
-    def merge(self, delta: ChoiceDeltaToolCall) -> None:
-        if delta.id:
-            self.id = delta.id
-        if delta.type:
-            self.type = delta.type
-        if delta.function is None:
-            return
-        if delta.function.name:
-            if self.name is None or self.name == delta.function.name:
-                self.name = delta.function.name
-            else:
-                self.name += delta.function.name
-        if delta.function.arguments:
-            self.arguments += delta.function.arguments
-
-    def as_tool_call(self, index: int) -> ChatCompletionMessageFunctionToolCall:
-        return ChatCompletionMessageFunctionToolCall(
-            id=self.id or f"call_{index}",
-            type=self.type or "function",
-            function=Function(name=self.name or "", arguments=self.arguments or "{}"),
-        )
-
-
-class _ToolCallAccumulator:
-    def __init__(self) -> None:
-        self._message_calls: list[ChatCompletionMessageToolCall] = []
-        self._stream_calls: dict[int, _StreamToolCall] = {}
-
-    def add_message_calls(self, calls: Iterable[ChatCompletionMessageToolCall]) -> None:
-        self._message_calls.extend(calls)
-
-    def merge_delta_calls(self, deltas: Iterable[ChoiceDeltaToolCall]) -> None:
-        for delta in deltas:
-            self._stream_calls.setdefault(delta.index, _StreamToolCall()).merge(delta)
-
-    def as_native(self) -> list[ChatCompletionMessageToolCall]:
-        if self._message_calls:
-            return list(self._message_calls)
-        return [self._stream_calls[index].as_tool_call(index) for index in sorted(self._stream_calls)]
-
-
-def _tool_invocation_from_native(
-    tool_call: ChatCompletionMessageToolCall,
-    tool_map: dict[str, Tool],
-) -> tuple[Tool, dict[str, Any]]:
-    tool_name, arguments = _parse_native_function_call(tool_call)
-    tool_obj = tool_map.get(tool_name)
-    if tool_obj is None:
-        raise BubError(ErrorKind.TOOL, f"Unknown tool name: {tool_name}.")
-    return tool_obj, arguments
-
-
-def _parse_native_function_call(tool_call: ChatCompletionMessageToolCall) -> tuple[str, dict[str, Any]]:
-    if not isinstance(tool_call, ChatCompletionMessageFunctionToolCall):
-        raise BubError(ErrorKind.INVALID_INPUT, "Expected a function tool call with JSON object arguments.")
-    try:
-        arguments = TOOL_ARGUMENTS_ADAPTER.validate_json(tool_call.function.arguments or "{}")
-    except ValidationError as exc:
-        raise BubError(ErrorKind.INVALID_INPUT, "Expected a function tool call with JSON object arguments.") from exc
-    return tool_call.function.name, arguments
-
-
-async def _completion_events(
-    completion: ChatCompletion | ParsedChatCompletion[Any] | AsyncIterator[ChatCompletionChunk],
-    state: StreamState,
-    text_parts: list[str],
-    tool_calls: _ToolCallAccumulator,
-) -> AsyncGenerator[StreamEvent, None]:
-    if isinstance(completion, ChatCompletion):
-        if usage := TapeService._extract_usage(completion):
-            state.usage = usage
-        message = completion.choices[0].message
-        for event in _completion_message_events(message, text_parts, tool_calls):
-            yield event
-        return
-
-    async for chunk in completion:
-        async for event in _completion_chunk_events(chunk, state, text_parts, tool_calls):
-            yield event
-
-
-def _completion_message_events(
-    message: ChatCompletionMessage,
-    text_parts: list[str],
-    tool_calls: _ToolCallAccumulator,
-) -> Iterable[StreamEvent]:
-    if message.reasoning:
-        yield StreamEvent("reasoning", {"delta": _reasoning_text(message.reasoning)})
-    if message.content:
-        text_parts.append(message.content)
-        yield StreamEvent("text", {"delta": message.content})
-    tool_calls.add_message_calls(cast("Iterable[ChatCompletionMessageToolCall]", message.tool_calls or []))
-
-
-async def _completion_chunk_events(
-    chunk: ChatCompletionChunk,
-    state: StreamState,
-    text_parts: list[str],
-    tool_calls: _ToolCallAccumulator,
-) -> AsyncGenerator[StreamEvent, None]:
-    if usage := TapeService._extract_usage(chunk):
-        state.usage = usage
-    for choice in chunk.choices:
-        delta = choice.delta
-        if delta.reasoning:
-            yield StreamEvent("reasoning", {"delta": _reasoning_text(delta.reasoning)})
-        if delta.content:
-            text_parts.append(delta.content)
-            yield StreamEvent("text", {"delta": delta.content})
-        if delta.tool_calls:
-            tool_calls.merge_delta_calls(delta.tool_calls)
-
-
-def _reasoning_text(reasoning: object) -> str:
-    content = getattr(reasoning, "content", reasoning)
-    return "" if content is None else str(content)
 
 
 @dataclass(frozen=True)
@@ -667,11 +494,6 @@ def _parse_args(args_tokens: list[str]) -> Args:
         else:
             positional.append(token)
     return Args(positional=positional, kwargs=kwargs)
-
-
-def _is_context_length_error(error_msg: str) -> bool:
-    """Check whether an error message indicates a context-length / prompt-too-long failure."""
-    return bool(_CONTEXT_LENGTH_PATTERNS.search(error_msg))
 
 
 def _extract_text_from_parts(parts: list[dict]) -> str:
