+++ tests/test_builtin_agent.py
@@ -9,6 +9,7 @@ import pytest
 from any_llm.types.completion import ChatCompletionChunk
 
 from bub.builtin.agent import Agent
+from bub.builtin.model_runner import ModelRunner
 from bub.builtin.settings import AgentSettings
 from bub.runtime import BubError
 from bub.tape import Tape, TapeContext
@@ -17,6 +18,16 @@ from bub.tools import REGISTRY, tool
 # ---------------------------------------------------------------------------
 # Agent.run() tests: merge_back logic and model passthrough
 # ---------------------------------------------------------------------------
+
+
+class _FakeModelRunner(ModelRunner):
+    def __init__(self, settings: AgentSettings) -> None:
+        super().__init__(settings)
+        self.completion_kwargs: dict[str, Any] | None = None
+
+    async def completion_response(self, **kwargs: Any) -> AsyncIterator[ChatCompletionChunk]:
+        self.completion_kwargs = kwargs
+        return _chat_stream("done")
 
 
 def _make_agent() -> Agent:
@@ -30,14 +41,13 @@ def _make_agent() -> Agent:
 
     agent.settings = AgentSettings.model_construct(model="test:model", api_key="k", api_base="b", client_args={})
     agent.framework = framework
-
-    async def fake_completion_response(**kwargs: Any) -> AsyncIterator[ChatCompletionChunk]:
-        agent.completion_kwargs = kwargs
-        return _chat_stream("done")
-
-    agent.completion_kwargs = None
-    agent._completion_response = fake_completion_response  # type: ignore[method-assign]
+    agent.model_runner = _FakeModelRunner(agent.settings)
     return agent
+
+
+def _model_runner(agent: Agent) -> _FakeModelRunner:
+    assert isinstance(agent.model_runner, _FakeModelRunner)
+    return agent.model_runner
 
 
 def _chat_chunk(content: str) -> ChatCompletionChunk:
@@ -186,13 +196,15 @@ async def test_agent_run_passes_model_to_llm() -> None:
     )
     [event async for event in result]
 
-    assert agent.completion_kwargs["model"] == "openai:gpt-4o"
+    completion_kwargs = _model_runner(agent).completion_kwargs
+    assert completion_kwargs is not None
+    assert completion_kwargs["model"] == "openai:gpt-4o"
 
 
 @pytest.mark.asyncio
 async def test_agent_run_empty_prompt_returns_error() -> None:
     agent = _make_agent()
-    agent.tapes = MagicMock()  # type: ignore[assignment]
+    agent.tapes = MagicMock()
 
     result = await agent.run_stream(session_id="user/s1", prompt="", state={})
     events = [event async for event in result]
@@ -214,7 +226,9 @@ async def test_agent_run_model_defaults_to_none() -> None:
     result = await agent.run_stream(session_id="user/s1", prompt="hello", state={"_runtime_workspace": "/tmp"})  # noqa: S108
     [event async for event in result]
 
-    assert agent.completion_kwargs["model"] == "test:model"
+    completion_kwargs = _model_runner(agent).completion_kwargs
+    assert completion_kwargs is not None
+    assert completion_kwargs["model"] == "test:model"
 
 
 @pytest.mark.asyncio
@@ -245,9 +259,10 @@ async def test_agent_run_resolves_allowed_tool_aliases_and_limits_prompt() -> No
     )
     [event async for event in result]
 
-    assert agent.completion_kwargs is not None
-    assert [tool.name for tool in agent.completion_kwargs["tools"]] == ["tests_allowed_agent_tool"]
-    system_prompt = agent.completion_kwargs["messages"][0]["content"]
+    completion_kwargs = _model_runner(agent).completion_kwargs
+    assert completion_kwargs is not None
+    assert [tool.name for tool in completion_kwargs["tools"]] == ["tests_allowed_agent_tool"]
+    system_prompt = completion_kwargs["messages"][0]["content"]
     assert "- tests_allowed_agent_tool(): Allowed tool" in system_prompt
     assert "tests_denied_agent_tool" not in system_prompt
